# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports/progress": {
      "post": {
        "operationId": "ReportController_generateReport",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateProgressInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressReportOutputDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Report"
        ]
      }
    },
...
    "/resources/events": {
      "post": {
        "operationId": "ResourceController_scheduleEvent",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleEventInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Resource"
        ]
      }
    },
    "/research/milestones": {
      "post": {
        "operationId": "ResearchController_uploadMilestone",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadMilestoneInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Research"
        ]
      }
    },
    "/research/milestones/{id}/review": {
      "patch": {
        "operationId": "ResearchController_reviewMilestone",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Milestone ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewMilestoneInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Research"
        ]
      }
    },
...
    "/leaves": {
      "post": {
        "operationId": "LeaveController_applyLeave",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyLeaveInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leave"
        ]
      }
    },
    "/leaves/{id}/approve": {
      "patch": {
        "operationId": "LeaveController_approveLeave",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Leave ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveLeaveInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leave"
        ]
      }
    },
...
    "/outcomes/map": {
      "post": {
        "operationId": "OutcomeController_mapAssessmentToOutcomes",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapOutcomeInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Outcome"
        ]
      }
    },
    "/outcomes/student/{studentId}": {
      "get": {
        "operationId": "OutcomeController_getStudentOutcomes",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentOutcomeOutputDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Outcome"
        ]
      }
    }
  },
  "info": {
    "title": "Academic Management System API",
    "description": "In-memory NestJS backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GenerateProgressInputDto": {
        "type": "object",
        "properties": {
          "faculty_id": {
            "type": "string",
            "description": "Faculty UUID",
            "example": "uuid-string"
          },
          "course_id": {
            "type": "string",
            "description": "Course UUID",
            "example": "uuid-string"
          }
        },
        "required": [
          "faculty_id"
        ]
      },
      "MarksDetailDto": {
        "type": "object",
        "properties": {
          "assessment": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "max": {
            "type": "number"
          }
        },
        "required": [
          "assessment",
          "score",
          "max"
        ]
      },
      "ProgressReportOutputDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "string"
          },
          "student_name": {
            "type": "string"
          },
          "marks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarksDetailDto"
            }
          },
          "internal_marks_placeholder": {
            "type": "number"
          },
          "graded_placeholder": {
            "type": "string"
          },
          "generated_at": {
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "student_name",
          "marks",
          "internal_marks_placeholder",
          "graded_placeholder",
          "generated_at"
        ]
      },
...
      "ScheduleEventInputDto": {
        "type": "object",
        "properties": {
          "resource_id": {
            "type": "string"
          },
          "start_time": {
            "type": "string"
          },
          "end_time": {
            "type": "string"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "lecture",
              "assessment",
              "seminar"
            ]
          }
        },
        "required": [
          "resource_id",
          "start_time",
          "end_time",
          "event_type"
        ]
      },
      "EventOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "resource_id": {
            "type": "string"
          },
          "start_time": {
            "type": "string"
          },
          "end_time": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "venue_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "resource_id",
          "start_time",
          "end_time",
          "event_type"
        ]
      },
      "UploadMilestoneInputDto": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "string"
          },
          "file_type": {
            "type": "string",
            "enum": [
              "PDF",
              "DOCX"
            ]
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "file_type",
          "description"
        ]
      },
      "MilestoneOutputDto": {
        "type": "object",
        "properties": {
          "milestone_id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "submission_date": {
            "type": "string"
          },
          "file_type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "milestone_id",
          "project_id",
          "status"
        ]
      },
      "ReviewMilestoneInputDto": {
        "type": "object",
        "properties": {
          "milestone_id": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED",
              "REVISION_NEEDED"
            ]
          }
        },
        "required": [
          "milestone_id",
          "comments",
          "status"
        ]
      },
...
      "ApplyLeaveInputDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "end_date": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "doc_ref": {
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "start_date",
          "end_date",
          "reason"
        ]
      },
      "LeaveOutputDto": {
        "type": "object",
        "properties": {
          "leave_id": {
            "type": "string"
          },
          "student_id": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "end_date": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "doc_ref": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "leave_id",
          "student_id",
          "start_date",
          "end_date",
          "reason",
          "status"
        ]
      },
      "ApproveLeaveInputDto": {
        "type": "object",
        "properties": {
          "leave_id": {
            "type": "string"
          }
        },
        "required": [
          "leave_id"
        ]
      },
...
      "MapOutcomeInputDto": {
        "type": "object",
        "properties": {
          "assessment_id": {
            "type": "string"
          },
          "outcome_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "weightages": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "assessment_id",
          "outcome_ids",
          "weightages"
        ]
      },
      "StudentOutcomeOutputDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "string"
          },
          "outcome_id": {
            "type": "string"
          },
          "outcome_title": {
            "type": "string"
          },
          "achievement_level": {
            "type": "string"
          },
          "raw_percentage": {
            "type": "number"
          }
        },
        "required": [
          "student_id",
          "outcome_id",
          "outcome_title",
          "achievement_level",
          "raw_percentage"
        ]
      }
    }
  }
}